requirements.txt
flask
flask-cors
pydantic
groq
python-dotenv
psycopg2-binary
gunicorn
